Many Fixes, wow
I don't really have a good commit message for this because I've been
"fixing" this as I was working on getting
https://github.com/freshbooks/charlesbot-pagerduty-escalations going.

I know, this is terrible :(
