+++ {{cookiecutter.repo_name}}/.gitignore
@@ -41,3 +41,7 @@ output/*/index.html
 
 # Sphinx
 docs/_build
+
+# Charlesbot
+development.yaml
+env/
