+++ {{cookiecutter.repo_name}}/Makefile
@@ -12,6 +12,13 @@ help:
 	@echo 'checkstyle: Run flake8'
 	@echo 'run: Run CharlesBOT locally'
 
+# Utility target for checking required parameters
+guard-%:
+	@if [ "$($*)" = '' ]; then \
+     echo "Missing required $* variable."; \
+     exit 1; \
+   fi;
+
 clean:
 	py3clean .
 	rm -f .coverage
@@ -41,3 +48,9 @@ test: install
 
 run:
 	PYTHONWARNINGS=default PYTHONASYNCIODEBUG=1 $(ENV)/bin/charlesbot
+
+# e.g. PART=major make release
+# e.g. PART=minor make release
+# e.g. PART=patch make release
+release: guard-PART
+  $(ENV)/bin/bumpversion $(PART)
