+++ {{cookiecutter.repo_name}}/setup.cfg
@@ -2,5 +2,6 @@
 current_version = {{ cookiecutter.version }}
 commit = True
 tag = True
+message = Update the {{ cookiecutter.repo_name }} version from {current_version} to {new_version}
 
 [bumpversion:file:setup.py]
